+++ src/Core/Query/AbstractRequestBuilder.php
@@ -21,6 +21,7 @@ abstract class AbstractRequestBuilder implements RequestBuilderInterface
     {
         $request = new Request();
         $request->setHandler($query->getHandler());
+        $request->addParam('distrib', $query->getDistrib());
         $request->addParam('omitHeader', $query->getOmitHeader());
         $request->addParam('timeAllowed', $query->getTimeAllowed());
         $request->addParam('NOW', $query->getNow());
@@ -29,7 +30,7 @@ abstract class AbstractRequestBuilder implements RequestBuilderInterface
 
         $request->addParam('wt', $query->getResponseWriter());
         if ($query::WT_JSON === $query->getResponseWriter()) {
-            // only one JSON format is supported
+            // Only flat JSON format is supported. Other JSON formats are easier to handle but might loose information.
             $request->addParam('json.nl', 'flat');
         }