+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [4.3.0-alpha.1]
 ### Added
 - Experimental support for collection API
+- Parameter 'distrib' for queries
 
 ### Changed
 - Deprecation of Endpoint::getBaseUri is revoked! It transparently forwards to getCollectionBaseUri or getCoreBaseUri now
@@ -17,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Zend 1.x support
 - PECL::Http adapter
 - PHP 7.0 support
+- Solr 1.4 result parser
 
 ### Fixed
 - Support for add-distinct and removeregex modifiers in Document::setFieldModifier